+++ bayes_opt/__init__.py
@@ -2,7 +2,7 @@ from .bayesian_optimization import BayesianOptimization, Events
 from .domain_reduction import SequentialDomainReductionTransformer
 from .util import UtilityFunction
 from .logger import ScreenLogger, JSONLogger
-from .advanced_constraint import ConstraintModel
+from .constraint import ConstraintModel
 
 __all__ = [
     "BayesianOptimization",
