+++ bayes_opt/constraint.py
